+++ src/dialogs/save_file_dialog.rs
@@ -149,12 +149,15 @@ where
         self.cancel.render(graphics, mouse_xy);
     }
 
-    fn on_key_press(&mut self, key: VirtualKeyCode, _: &Vec<&VirtualKeyCode>) {
+    fn on_key_up(&mut self, key: VirtualKeyCode, _: &Vec<&VirtualKeyCode>) {
         self.name_field.on_key_press(key);
         self.current_dir_field.on_key_press(key);
     }
 
-    fn on_mouse_click(&mut self, xy: Coord, _: &Vec<&VirtualKeyCode>) {
+    fn on_mouse_up(&mut self, xy: Coord, button: MouseButton, _: &Vec<&VirtualKeyCode>) {
+        if button != MouseButton::Left {
+            return;
+        }
         if self.cancel.on_mouse_click(xy) {
             self.result = SceneUpdateResult::Pop(None);
         }
@@ -212,7 +215,7 @@ where
         }
     }
 
-    fn on_scroll(&mut self, xy: Coord, y_diff: isize, _: isize) {
+    fn on_scroll(&mut self, xy: Coord, _: isize, y_diff: isize, _: &Vec<&VirtualKeyCode>) {
         self.dir_panel.on_scroll(xy, y_diff);
     }